split events
