+++ spec/datadog/core/telemetry/event/distributions_spec.rb
@@ -1,0 +1,32 @@
+require 'spec_helper'
+
+require 'datadog/core/telemetry/event'
+require 'datadog/core/telemetry/metric'
+
+RSpec.describe Datadog::Core::Telemetry::Event::Distributions do
+  let(:id) { double('seq_id') }
+  let(:event) { event_class.new }
+
+  subject(:payload) { event.payload }
+
+  let(:event_class) { described_class }
+  let(:event) { event_class.new(namespace, metrics) }
+
+  let(:namespace) { 'general' }
+  let(:metric_name) { 'request_duration' }
+  let(:metric) do
+    Datadog::Core::Telemetry::Metric::Distribution.new(metric_name, tags: { status: '200' })
+  end
+  let(:metrics) { [metric] }
+
+  let(:expected_metric_series) { [metric.to_h] }
+
+  it do
+    is_expected.to eq(
+      {
+        namespace: namespace,
+        series: expected_metric_series
+      }
+    )
+  end
+end
